Implement checkpoint cleanup for leader mode (#1076)

+++ crates/arroyo-connectors/src/filesystem/sink/v2/mod.rs
@@ -835,7 +835,7 @@ impl<BBW: BatchBufferingWriter + Send + 'static> ArrowOperator for FileSystemSin
         // Send completion event
         ctx.control_tx
             .send(arroyo_rpc::ControlResp::CheckpointEvent(CheckpointEvent {
-                checkpoint_epoch: epoch,
+                checkpoint_epoch: epoch as u64,
                 operator_idx: ctx.task_info.operator_idx,
                 operator_id: ctx.task_info.operator_id.clone(),
                 subtask_idx: ctx.task_info.task_index,
