+++ crates/arroyo-connectors/src/kafka/sink/mod.rs
@@ -378,7 +378,7 @@ impl ArrowOperator for KafkaSinkFunc {
             }
         }
         let checkpoint_event = ControlResp::CheckpointEvent(CheckpointEvent {
-            checkpoint_epoch: epoch,
+            checkpoint_epoch: epoch as u64,
             operator_idx: ctx.task_info.operator_idx,
             operator_id: ctx.task_info.operator_id.clone(),
             subtask_idx: ctx.task_info.task_index,
